auto mode inserted
